fix: Headerをコンポーネントとして切り出し #21
